fix windows dev server bin resolve
