+++ examples/nextjs/vercel.json
@@ -1,7 +1,7 @@
 {
   "functions": {
     "api/**/*.rs": {
-      "runtime": "vercel-rust@4.0.0-beta.1"
+      "runtime": "vercel-rust@4.0.0-beta.2"
     }
   }
 }
