+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "vercel-rust",
-  "version": "4.0.0-beta.1",
+  "version": "4.0.0-beta.2",
   "description": "Rust runtime for Vercel Functions.",
   "homepage": "https://github.com/vercel-community/rust",
   "repository": {
